.claude/worktrees/agent-a4cfb9a6/requirements.txt
# VÉLØ v9.0++ CHAREX - Oracle of Odds
# Python Dependencies

# Core
requests>=2.31.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# API Integration
flask>=3.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.7,<3.0
pydantic-settings>=2.0,<3.0
orjson>=3.9.0

# Database & Storage
pymongo>=4.5.0
supabase>=2.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Analysis & ML
scikit-learn>=1.3.0
scipy>=1.11.0
lightgbm>=4.0.0
joblib>=1.3.0

# Feature Store & Monitoring
feast>=0.38.0
evidently>=0.4.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Logging & Monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.1.0
ruff>=0.1.0
mypy>=1.5.0

